+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Updated favicon.
 - Fixed "edit" button in Review page for the Address fields.
 - Updated the list of library names.
+- Fixed issue rendering non-object error messages in the `ApiErrors` component.
 
 ### v0.6.0